NSUInteger / NSInteger transitions for count/length variables.
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/apps/projectcenter/trunk@37135 72102866-910b-0410-8b05-ffd578937521

+++ Modules/Parsers/ProjectCenter/ObjCClassHandler.m
@@ -93,7 +93,7 @@
 
 - (void)string:(NSString *)element
 {
-  unsigned int len = [element length];
+  NSUInteger len = [element length];
 
   [super string:element];
